Double precision

+++ src/deform/registration/settings.h
@@ -52,7 +52,7 @@ struct Settings
     // Block size, (0,0,0) is the same as using only a single large block
     int3 block_size;
     // Epsilon used for termination
-    float block_energy_epsilon;
+    double block_energy_epsilon;
     // Step size in [mm]
     float step_size;
     // Only considered if no weight map is given
